refactor: 宏降低两种取最高负载方案的冗余
